Initialized the home page w/ django
Everything is kept inside of the "PriceChecker" folder, this acts as the root/main folder for the whole project. The name of this folder can change. For right now, when you start the server with django, it will prompt you to the homepage which only contains a simple h1 body. Note: Going to "home/" also does the same as having nothing after the port number.